# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SinoGlobal Enterprise API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsLeaveRequestsByStatusItem } from './dashboardStatsLeaveRequestsByStatusItem';
import type { DashboardStatsWorkersByDepartmentItem } from './dashboardStatsWorkersByDepartmentItem';
import type { Worker } from './worker';

export interface DashboardStats {
  totalWorkers: number;
  activeWorkers: number;
  pendingLeaveRequests: number;
  totalLeaveRequests: number;
  workersByDepartment: DashboardStatsWorkersByDepartmentItem[];
  leaveRequestsByStatus: DashboardStatsLeaveRequestsByStatusItem[];
  recentWorkers: Worker[];
}