# lib/api-zod/src/generated/types/jobRoleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SinoGlobal Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface JobRoleUpdate {
  title?: string;
  department?: string;
  description?: string;
  responsibilities?: string;
  requirements?: string;
  salary?: string;
  photoUrl?: string;
  isActive?: boolean;
}