# lib/api-zod/src/generated/types/leaveLetter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SinoGlobal Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveLetter {
  id: number;
  workerId: number;
  /** @nullable */
  workerName?: string | null;
  leaveTypeId: number;
  /** @nullable */
  leaveTypeName?: string | null;
  subject: string;
  body: string;
  recipientTitle: string;
  status?: string;
  createdAt: string;
}