# lib/api-zod/src/generated/types/leaveRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SinoGlobal Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveRequest {
  id: number;
  workerId: number;
  /** @nullable */
  workerName?: string | null;
  leaveTypeId: number;
  /** @nullable */
  leaveTypeName?: string | null;
  startDate: string;
  endDate: string;
  /** @nullable */
  reason?: string | null;
  /** pending | approved | rejected */
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}