# lib/api-zod/src/generated/types/leaveRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SinoGlobal Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveRequestInput {
  workerId: number;
  leaveTypeId: number;
  startDate: string;
  endDate: string;
  reason?: string;
}